Add initial version of 'Replace in project (rip)' plugin
This commit adds the initial version of the replace plugin, which provides users with a simple interface to replace occurrences of a given string throughout a given project. It includes a README file and a LICENSE file.

The plugin supports replacing in all files of the project and in all git tracked files. Users can also customize keybindings inside the replace window and how they start the plugin.

This commit squashes all the changes made to implement the above features.